\docType{package}
\name{gcomputation}
\alias{gcomputation}
\alias{gcomputation-package}

\title{
gcomputation: Estimating Marginal Effects Using G-Computation
}

\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='gcomputation logo' width='120'}}

The \pkg{gcomputation} package provides functions to compute
G-Computation (GC) to estimate marginal effects. It includes functions
for estimating marginal effects for binary, time-to-event, continuous
and count outcomes regarding two exposures. The package implements GC
with various working models or algorithms, referred to as Q-models.
}

\section{Key Features}{

The functions \code{gc_binary()}, \code{gc_times()},
\code{gc_continuous()} and \code{gc_count()} implement
G-Computation to estimate marginal functions for binary,
time-to-event, continuous and count outcomes regarding two
exposures using a variety of modeling strategies.

The package supports several methods to construct the Q-model:

\itemize{
\item \code{"all"}: Uses a usual logistic or Cox model, incorporating all variables provided in the formula.
\item \code{"lasso"}: Implements L1 regularization regression, which can perform predictor selection. It uses the \pkg{glmnet} package.
\item \code{"ridge"}: Applies L2 regularization regression, also utilizing the \pkg{glmnet} package. This is equivalent to Elastic Net with an alpha value of 0.
\item \code{"elasticnet"}: Combines both L1 and L2 regularization regression, also using the \pkg{glmnet} package. The alpha parameter controls the mix between L1 and L2, typically ranging from 0 to 1.
\item \code{"aic"}: Performs forward selection regression based on the Akaike Information Criterion (AIC), using \code{stepAIC()}.
\item \code{"bic"}: Performs forward selection regression based on the Bayesian Information Criterion (BIC), also using \code{stepAIC()} with \code{k = log(nrow(data))}.
}

The package offers estimation of three types of marginal effects:

\itemize{
\item \code{"ATE"} (Average Treatment Effect on the Entire Population): The marginal effect if the entire sample were treated versus entirely untreated.
\item \code{"ATT"} (Average Treatment Effect on the Treated): The marginal effect if the treated patients (group = 1) would have been untreated.
\item \code{"ATU"} (Average Treatment Effect on the Untreated): The marginal effect if the untreated patients (group = 0) would have been treated.
}

S3 methods are included for objects generated by
\code{gc_binary()} and \code{gc_times()}, allowing for:

\itemize{
\item \code{print()}: Print a summary of the results.
\item \code{summary()}: Provide a more detailed summary of the prognostic capacities, including confidence intervals.
\item \code{plot()}: Visualize the results through calibration plots or effect-specific plots (proportion for logistic, survival curve for survival).
}

Other exported functions and data:

\itemize{
\item \code{transport()}: Applies an already fitted GC model (an object of class \code{gcbinary}, \code{gctimes}, \code{gccontinuous} or \code{gccount}) to a \code{newdata} set to estimate marginal effects in a new population.
\item Datasets: \code{dataPROPHYVAP} (simulated randomized clinical trial data) and \code{dataCOHORT} (simulated observational cohort data).
\item Multiple Imputations (MI-BOOT): Support for the MI-BOOT approach is added using the \code{boot.mi} argument, integrating the \pkg{mice} package for handling missing data.
}

The package also supports bootstrapping for confidence interval estimation,
with options for \code{"bcv"} (default) or \code{"boot"} and a default of
500 bootstrap resamples. Users can also control whether tuning parameters
are estimated within each bootstrap iteration or on the total population.

}

\section{Basic Usage}{

For a binary outcome:

\preformatted{
data("dataPROPHYVAP")

.f <- formula(VAP ~ GROUP * (AGE + SEX + BMI + DIABETES))

# 1. Standard execution
# boot.tune=TRUE estimates tuning parameters inside each bootstrap iteration.
gc_bin <- gc_binary(formula=.f, model="ridge", data=dataPROPHYVAP,
                    group="GROUP", cv=10, boot.type="bcv",
                    boot.number=500, boot.tune=TRUE,
                    effect="ATE", progress=TRUE, seed=5192)
gc_bin

# Summary specifying Asymptotic CIs ("norm")
summary(gc_bin, ci.type="norm")

# Calibration plot
plot(gc_bin)

# 2. Execution with multiple imputation
# Uses boot.mi=TRUE and m=5. boot.tune=FALSE to only estimate it once on the complete data set
.f_mi <- formula(VAP ~ GROUP * (AGE + SEX + BMI + DIABETES + GLASGOW + INJURY))
gc_mi <- gc_binary(formula=.f_mi, model="elasticnet",
                   data=dataPROPHYVAP, group="GROUP",
                   cv=10, boot.type="bcv",
                   boot.number=500, boot.tune=FALSE,
                   effect="ATE", progress=TRUE,
                   seed=8051, boot.mi=TRUE, m=5)

# Plotting the calibration curve, smoothed across m imputations
plot(gc_mi, smooth=TRUE)

# Summary specifying Non-parametric CIs ("perc")
summary(gc_mi, ci.type="perc")

# 3. Transportability
# Define a new dataset (e.g., a subset of younger patients, AGE<=50)
newdata_binary <- subset(dataPROPHYVAP, AGE <= 50)

# Transport the fitted gc_bin model to the new dataset
gc_transport <- transport(object=gc_bin,
                          newdata=newdata_binary,
                          boot.number=500)

summary(gc_transport, ci.type="norm")
}

For a survival outcome:

\preformatted{
data(dataPROPHYVAP)

.ft <- formula(Surv(TIME_DEATH, DEATH) ~ GROUP * (AGE + BMI + GLASGOW + LEUKO))

gc_surv <- gc_times(formula=.ft, model="lasso",
                    data=dataPROPHYVAP, group="GROUP",
                    param.tune=0.03,
                    boot.type="bcv", boot.number=500,
                    boot.tune=FALSE,
                    effect="ATE",
                    pro.time=30,
                    seed=5312)

gc_surv
summary(gc_surv, ci.type="perc")

# Predicted survival curves for treatment groups
plot(gc_surv, method="survival", col=c("red3","blue3"))
legend("bottomleft",
       c("Placebo", "Ceftriaxone"),
       col=c("red3","blue3"),
       lty=1)
}

}

\seealso{
Useful links:

\itemize{
\item GitHub repository: \url{https://github.com/chupverse/gcomputation}
\item Report bugs at \url{https://github.com/chupverse/gcomputation/issues}
}
}

\author{

\strong{Maintainer:}

Yohann Foucher
\email{yohann.foucher@univ-poitiers.fr}
(\href{https://orcid.org/0000-0003-0330-7457}{ORCID})

\strong{Authors:}

\itemize{
\item Joe De Keizer
\email{joe.de.keizer@univ-poitiers.fr}
(\href{https://orcid.org/0000-0003-0821-4540}{ORCID})
}

}

\keyword{package}