\name{miboot_gc_survival}
\alias{miboot_gc_survival}
\title{
G-computation for Marginal Survival Effects with Multiple Imputation
}
\description{
This function performs G-computation (GC) for marginal survival effects on datasets with missing values by first applying multiple imputation using the \code{\link[mice]{mice}} package and then running \code{\link{gc_survival}} on each imputed dataset. The bootstrap samples from all imputed datasets are combined to provide a single set of estimates.
}
\usage{
miboot_gc_survival(formula, data, group, pro.time, ..., m = 5,
                    effect = "ATE", gc.method, param.tune = NULL, cv = 10,
                    boot.type = "bcv", boot.number = 500,
                    boot.tune = FALSE, progress = TRUE, seed = NULL)
}
\arguments{
  \item{formula}{
A survival formula related to the Q-model with the variable \code{group} among the predictors. For details, see \code{\link{gc_survival}}.
}
  \item{data}{
A data frame containing the variables for the outcome (time and event), the exposure (\code{group}), and the predictors. This dataset may contain missing values, which will be imputed by multiple imputation.
}
  \item{group}{
The name of the variable related to the exposure/treatment. This variable should have two modalities (0 for untreated/unexposed and 1 for treated/exposed).
}
  \item{pro.time}{
An optional value for censoring the follow-up times and obtaining survival curves and related restricted mean survival times up to \code{pro.time}. The default value is median of the follow-up times. For details, see \code{\link{gc_survival}}.
}
  \item{...}{
Additional arguments to be passed directly to the \code{\link[mice]{mice}} function for customizing the multiple imputation process (e.g., \code{method}, \code{maxit}, \code{diagnostics}).
}
  \item{m}{
The number of imputed datasets to create using \code{mice}. The default value is 5.
}
  \item{effect}{
The type of marginal effect to be estimated ("ATE", "ATT", or "ATU"). The default is "ATE". For details, see \code{\link{gc_survival}}.
}
  \item{gc.method}{
The method used to create the Q-model for G-computation (e.g., "all", "lasso", "ridge", "elasticnet", "aic", "bic"). This argument is named \code{gc.method} here to avoid conflict with the \code{method} argument in the \code{\link[mice]{mice}} function. For details, see \code{method} in \code{\link{gc_survival}}.
}
  \item{param.tune}{
An optional argument to specify the tuning parameters for the Q-model. If \code{NULL} (the default), tuning parameters are estimated by cross-validation for each algorithm.
}
  \item{cv}{
The number of splits for cross-validation used during Q-model fitting. The default value is 10.
}
  \item{boot.type}{
The type of bootstrap to use ("bcv" or "boot"). The default value is "bcv".
}
  \item{boot.number}{
The number of bootstrap resamples to perform *for each imputed dataset*. The total number of bootstrap samples in the final combined results will be \code{m} * \code{boot.number}. The default value is 500.
}
  \item{boot.tune}{
A logical value to determine whether the tuning parameter should be estimated inside of each bootstrap iteration for \code{gc_survival}.
}
  \item{progress}{
A logical value to print a progress bar for the imputation process and the G-computation loop in the R console. The default is \code{TRUE}.
}
  \item{seed}{
A random seed to ensure reproducibility of the imputation process and the bootstrapping steps within \code{gc_survival}. If \code{NULL} (the default), a seed is randomly assigned.
}
}
\details{
The \code{miboot_gc_survival} function integrates multiple imputation for handling missing data with G-computation for causal inference. The steps it does:
\enumerate{
    \item Multiple Imputation: The input \code{data} is first subjected to multiple imputation using the \code{\link[mice]{mice}} function. This creates \code{m} complete datasets. Additional arguments for customizing the \code{mice} function (e.g., imputation methods for different variable types) can be passed via the \code{...} argument.
    \item G-computation and combination: For each of the \code{m} imputed datasets, the \code{\link{gc_survival}} function is called to estimate the marginal survival effects. All parameters specific to \code{gc_survival} are passed directly. The raw bootstrap estimates (e.g., Average Hazard Ratios, Restricted Mean Survival Times, survival probabilities) generated by each \code{gc_survival} call are concatenated. This results in a final set of bootstrap samples that accounts for both imputation uncertainty and sampling variability, following the MI BOOT method (see references).
}

The final returned object is structured similarly to a standard \code{gcsurv} object, ensuring compatibility with its S3 methods (\code{print.gcsurv}, \code{summary.gcsurv}). However, several fields are modified by the multiple imputation process:
\itemize{
    \item \code{calibration}: This field contains a list of the \code{calibration} objects (each with a fitted Q-model, predicted time points, cumulative hazards, survival probabilities, baseline cumulative hazard, and linear predictors) from each of the \code{m} \code{gc_survival} runs.
    \item \code{tuning.parameters}: This field contains a list of vectors (\code{lambda} and \code{alpha}) which store all the optimal tuning parameters identified across the \code{m} \code{gc_survival} runs.
    \item \code{data}: This field contains a list of the \code{m} complete imputed datasets.
    \item \code{initial.data}: This new field stores the original \code{data} frame as supplied to \code{miboot_gc_survival}, before any imputation with the missing data.
    \item \code{boot.number}: This field reflects the total number of combined bootstrap samples (\code{m} * \code{boot.number} argument).
    \item \code{m}: This new field indicates the number of imputed datasets used from the initial call.
    \item \code{pro.time}: This field specifies the common time point up to which RMST and survival probabilities are estimated across all imputations.
}
}
\value{
The function returns an object of class \code{"gcsurv"}, which is a list containing the following elements, updated to reflect the multiple imputation and aggregation:
\item{calibration}{A list of \code{m} lists, where each inner list contains the final fitted Q-model (\code{fit}), predicted time points (\code{time}), cumulative hazards (\code{cumhaz}), survival probabilities (\code{surv}), baseline cumulative hazard (\code{H0.multi}), and linear predictors (\code{lp}) from one of the imputed datasets.}
\item{tuning.parameters}{A list containing vectors of optimal \code{lambda} and \code{alpha} values (if applicable), where each vector element corresponds to the optimal parameter from one of the \code{m} imputed datasets.}
\item{data}{A list of \code{m} data frames, where each element is one of the complete imputed datasets used in the G-computation.}
\item{formula}{The original formula provided to the function.}
\item{method}{The method used for the Q-model (e.g., "lasso", "ridge", "aic").}
\item{cv}{The number of splits used for cross-validation for the Q-model.}
\item{missing}{The number of observations that were removed due to missing values during the processing of the *first* imputed dataset (which defines the structural elements of the resulting \code{gcsurv} object).}
\item{pro.time}{The specific time point up to which RMST and survival probabilities are estimated.}
\item{boot.number}{The total number of bootstrap resamples performed across all imputed datasets (\code{m} * \code{boot.number} argument).}
\item{boot.type}{The type of bootstrap resamples performed.}
\item{outcome}{A list of character strings, specifying the names of the time (\code{times}) and outcome variable.}
\item{group}{A character string specifying the name of the grouping (exposure/treatment) variable.}
\item{n}{The sample size of the imputed dataset used in calculations.}
\item{nevent}{The total number of events (outcome coded as 1) in the imputed dataset used in calculations.}
\item{AHR}{A numeric vector containing the raw bootstrap samples of the estimated Average Hazard Ratio, combined from all \code{m} imputed datasets.}
\item{RMST0}{A numeric vector containing the raw bootstrap samples of the estimated Restricted Mean Survival Time for the control group, up to \code{pro.time}, combined from all \code{m} imputed datasets.}
\item{RMST1}{A numeric vector containing the raw bootstrap samples of the estimated Restricted Mean Survival Time for the experimental group, up to \code{pro.time}, combined from all \code{m} imputed datasets.}
\item{deltaRMST}{A numeric vector containing the raw bootstrap samples of the estimated difference in Restricted Mean Survival Times (\code{RMST1} - \code{RMST0}), combined from all \code{m} imputed datasets.}
\item{surv0}{A numeric vector containing the raw bootstrap samples of the estimated survival probability for the control group, at \code{pro.time}, combined from all \code{m} imputed datasets.}
\item{surv1}{A numeric vector containing the raw bootstrap samples of the estimated survival probability for the experimental group, at \code{pro.time}, combined from all \code{m} imputed datasets.}
\item{deltasurv}{A numeric vector containing the raw bootstrap samples of the estimated difference in survival probabilities (\code{surv1} - \code{surv0}), at \code{pro.time}, combined from all \code{m} imputed datasets.}
\item{AHR.unadj}{A numeric vector containing the raw bootstrap samples of the unadjusted Average Hazard Ratio, combined from all \code{m} imputed datasets.}
\item{RMST0.unadj}{A numeric vector containing the raw bootstrap samples of the unadjusted Restricted Mean Survival Time for the control group, up to \code{pro.time}, combined from all \code{m} imputed datasets.}
\item{RMST1.unadj}{A numeric vector containing the raw bootstrap samples of the unadjusted Restricted Mean Survival Time for the experimental group, up to \code{pro.time}, combined from all \code{m} imputed datasets.}
\item{deltaRMST.unadj}{A numeric vector containing the raw bootstrap samples of the unadjusted difference in Restricted Mean Survival Times (\code{RMST1.unadj} - \code{RMST0.unadj}), combined from all \code{m} imputed datasets.}
\item{surv0.unadj}{A numeric vector containing the raw bootstrap samples of the unadjusted survival probability for the control group, at \code{pro.time}, combined from all \code{m} imputed datasets.}
\item{surv1.unadj}{A numeric vector containing the raw bootstrap samples of the unadjusted survival probability for the experimental group, at \code{pro.time}, combined from all \code{m} imputed datasets.}
\item{deltasurv.unadj}{A numeric vector containing the raw bootstrap samples of the unadjusted difference in survival probabilities (\code{surv1.unadj} - \code{surv0.unadj}), at \code{pro.time}, combined from all \code{m} imputed datasets.}
\item{call}{The function call that generated the \code{miboot_gc_survival} object.}
\item{m}{The number of imputed datasets used.}
\item{initial.data}{The original data frame supplied to the \code{miboot_gc_survival} function before imputation.}
\item{seed}{The random seed used for the imputation process and for generating seeds for each \code{gc_survival} call.}
}
\references{
Chatton et al. G-computation and doubly robust standardisation for continuous-time data: A comparison with inverse probability weighting. Stat Methods Med Res. 31(4):706-718. 2022. <doi: 10.1177/09622802211047345>. 

Schomaker M, Heumann C. Bootstrap Inference When Using Multiple Imputation. Statistics in medicine. 2018;37(14):2252-2266. doi: 10.1002/sim.7654.

Van Buuren, S., & Groothuis-Oudshoorn, K. (2011). mice: Multivariate Imputation by Chained Equations in R. Journal of Statistical Software, 45(3), 1-67.
}
\examples{
data("dataPROPHYVAP")

dataPROPHYVAP$DEATH_num <- ifelse(dataPROPHYVAP$DEATH == "Yes",1,0)
dataPROPHYVAP$GROUP_num <- ifelse(dataPROPHYVAP$GROUP == "Placebo",0,1)

.f_surv <- formula(survival::Surv(TIME_DEATH, DEATH_num) ~ GROUP_num * (AGE + SEX + BMI + DIABETES))

#Note if it takes too long, reduce boot.number
gc1 <- miboot_gc_survival(formula=.f_surv, m=5, gc.method="ridge", data=dataPROPHYVAP,
                          group="GROUP_num", pro.time = 60, param.tune=NULL, boot.type="bcv", cv=10,
                          boot.number=500, effect="ATE", progress=TRUE, boot.tune = TRUE, seed = 42)

print(gc1)
summary(gc1)
plot(gc1) #creates m plots
}
\keyword{Causal inference}
\keyword{G-computation}
\keyword{Multiple imputation}
\keyword{Survival analysis}