\name{plot.gccount}
\alias{plot.gccount}
\title{
S3 Method for Plotting a 'gccontinuous' Object
}
\description{
Provides graphical summaries for an object returned by the function \code{gc_count}. This S3 method offers two types of plots for calibration or means' representation.
}
\usage{
\method{plot}{gccount}(x, method="calibration", n.groups=5, smooth=FALSE, ...)
}
\arguments{

\item{x}{An object returned by the function \code{gc_continuous}.}

\item{method}{A character string specifying the type of plot. Possible values are "calibration" (default) or "survival". See details.}

\item{n.groups}{An integer for the number of groups when \code{method="calibration"} is used. The default is 5.}

\item{smooth}{A logical value by default equals to \code{FALSE}. If \code{boot.mi=TRUE} in the object \code{x}, it enables plotting a single smooth calibration curve averaged over all \code{m} values instead of separate curves.}

\item{...}{Additional graphical parameters that can be passed to the underlying plot function.}
}
\details{
The \code{plot.gccount} function is an S3 method for visualizing results from an object generated by the \code{gc_count} function (i.e., G-Computation to estimate marginal functions for counting outcomes for two exposures). It supports two distinct types of visualizations:
\itemize{
\item \code{method="proportion"}: This plot compares the means of the outcome's predicted values for the non-exposed/untreated group and the exposed/treated group against the observed outcome's values. This provides a direct visual comparison of how well the Q-model average predictions align with the observed outcome's values within each exposure group.

\item \code{method="calibration"}: This plot assesses the calibration of the Q-model used within G-Computation.
    \itemize{
    \item It utilizes the predicted values of the outcome from the Q-model.
    \item These predicted values are then divided into \code{n.groups} (defaulting to 5) by their quantiles.
    \item For each strata, the mean predicted value is plotted  against the observed mean value of events.
    \item 95\% confidence intervals of these observed values are also displayed as vertical segments.
    }
}
}
\value{No return value for this S3 method.}
\examples{
data("dataPROPHYVAP")

.f <- formula(VAP ~ GROUP * (AGE + SEX + ALCOHOL + BMI + DIABETES))

### In practice use larger values of boot.number (e.g., 500).
### Here we set boot.number = 10 and cv = 10 for speed in CRAN checks.
### cv=30 and boot.number=1000 are more appropriate values
gc1 <- gc_count(formula=.f, model="all", data=dataPROPHYVAP, group="GROUP",
                    cv=10, boot.type="boot", boot.number=10, boot.tune=TRUE,
                    effect="ATE", progress=TRUE, seed=5192)

#Plot the calibration curve (by default)
plot(gc1, n.groups=5)
}

\keyword{Plotting}
\keyword{Counting outcome}