\name{plot.gclogi}
\alias{plot.gclogi}
\title{

S3 Method for Plotting an 'gclogi' Object

}
\description{

Provides graphical summaries for an object returned by the function \code{gc_logistic}.
This S3 method offers two types of plots: a calibration plot and a proportion plot.

}
\usage{

\method{plot}{gclogi}(x, method="calibration", n.groups=5, pro.time=NULL, ...)

}
\arguments{

\item{x}{An object returned by the function \code{gc_logistic}.}

\item{method}{A character string specifying the type of plot. Possible values are \code{calibration} (default) or \code{proportion}.}

\item{n.groups}{An integer for the number of groups to divide the predicted probabilities into when \code{method="calibration"} is used. The default is 5.}

\item{pro.time}{This argument is included for consistency with related S3 methods but is not applicable or used for logistic models.}

\item{...}{Additional graphical parameters that can be passed to the underlying plot function.}

}
\details{

The \code{plot.gclogi} function is an S3 method for visualizing results from a \code{gclogi} object, which is generated by the \code{gc_logistic} function. The \code{gc_logistic} function computes G-Computation to estimate marginal logistic functions for two exposures.

This plotting method supports two distinct types of visualizations:
\itemize{
\item \code{method="proportion"}: This plot compares the mean predicted probability of the event for the non-exposed/untreated group (0) and the exposed/treated group (1) against their observed proportions in the original dataset. This provides a direct visual comparison of how well the G-computation model's average predictions align with the observed event rates within each exposure group.

\item \code{method="calibration"}: This plot assesses the calibration of the Q-model's predictions used within G-Computation.
    \itemize{
    \item It utilizes the predicted probabilities of the event from the Q-model.
    \item These predicted probabilities are then divided into \code{n.groups} (defaulting to 5) by their quantiles, creating different risk strata.
    \item For each group, the mean predicted probability is plotted on the X axe against the observed proportion of events for that group on the Y axe.
    \item 95\% confidence intervals for these observed proportions are also displayed as vertical segments.
    \item An identity line (y=x) is included for reference. Perfect calibration is indicated when observed points are along this line, suggesting that the predicted probabilities accurately reflect the observed event rates across different risk groups.
    }
}

}
\value{No return value for this S3 method.}
\examples{
data(dataPROPHYVAP)

dataPROPHYVAP$VAP_EARLY_num <- ifelse(dataPROPHYVAP$VAP_EARLY == "Yes", 1, 0)
dataPROPHYVAP$GROUP_num <- ifelse(dataPROPHYVAP$GROUP == "Placebo", 0, 1)

formula <- formula(VAP_EARLY_num ~ GROUP_num * (AGE + SEXE + BMI + DIABETES))

gc1 <- gc_logistic(formula=formula, method="ridge", data=dataPROPHYVAP,
              group="GROUP_num", param.tune=NULL, boot.type="bcv", cv=10,
              boot.number=500, effect="ATE", progress=TRUE , boot.tune = TRUE)

#Plot the calibration curve (by default)
plot(gc1, method="calibration", n.groups=5)
	 
#Plot the proportion comparison
plot(gc1, method="proportion")
}

\keyword{Plotting}
\keyword{G-Computation}
\keyword{Logistic}