\name{plot.gccontinuous}
\alias{plot.gccontinuous}
\title{

S3 Method for Plotting an 'gccontinuous' Object

}
\description{

Provides graphical summaries for an object returned by the function \code{gc_continuous}.
This S3 method offers two types of plots: a calibration plot and a proportion plot.

}
\usage{

\method{plot}{gccontinuous}(x, method="calibration", n.groups=5, smooth=FALSE, ...)

}
\arguments{

\item{x}{An object returned by the function \code{gc_continuous}.}

\item{method}{A character string specifying the type of plot. Possible values are \code{calibration} (default) or \code{proportion}.}

\item{n.groups}{An integer for the number of groups to divide the predicted means into when \code{method="calibration"} is used. The default is 5. 
Note: If \code{n.groups} is set too high, some groups may have too few events to reliably estimate means. An error can also occur if the Q-model is singular.}

\item{smooth}{A logical value by default equal to \code{FALSE}. If \code{boot.mi} is TRUE, enables plotting a single smooth calibration curve averaged over all \code{m} values instead of separate curves.}

\item{...}{Additional graphical parameters that can be passed to the underlying plot function.}

}
\details{

The \code{plot.gccontinuous} function is an S3 method for visualizing results from a \code{gccontinuous} object, which is generated by the \code{gc_continuous} function. The \code{gc_continuous} function computes G-Computation to estimate marginal functions for continuous outcomes for two exposures.

This plotting method supports two distinct types of visualizations:
\itemize{
\item \code{method="proportion"}: This plot compares the means of the outcome's predicted values for the non-exposed/untreated group (0) and the exposed/treated group (1) against their observed outcome's values in the original dataset. This provides a direct visual comparison of how well the G-computation model's average predictions align with the observed outcome's values within each exposure group.

\item \code{method="calibration"}: This plot assesses the calibration of the Q-model's predictions used within G-Computation.
    \itemize{
    \item It utilizes the predicted values of the outcome from the Q-model.
    \item These predicted values are then divided into \code{n.groups} (defaulting to 5) by their quantiles, creating different strata.
    \item For each group, the mean predicted value is plotted on the X axe against the observed mean value of events for that group on the Y axe.
    \item 95\% confidence intervals for these observed values are also displayed as vertical segments.
    \item An identity line (y=x) is included for reference. Perfect calibration is indicated when observed points are along this line, suggesting that the predicted values accurately reflect the observed values across different risk groups.
    }
}

}
\value{No return value for this S3 method.}
\examples{
data("dataPROPHYVAP")

dataPROPHYVAP$VAP_num <- ifelse(dataPROPHYVAP$VAP == "Yes",1,0)
dataPROPHYVAP$GROUP_num <- ifelse(dataPROPHYVAP$GROUP == "Placebo",0,1)

beta_ttt <- 10
beta_alcool <- 2
beta_bmi <- 8
beta_age <- 5

set.seed(123)

cont_outcome <- rnorm(n = nrow(dataPROPHYVAP), 
                      mean = beta_ttt*dataPROPHYVAP$GROUP_num +
                        beta_alcool*match(as.factor(dataPROPHYVAP$ALCOOL),levels(as.factor(dataPROPHYVAP$ALCOOL))) + 
                        beta_bmi*dataPROPHYVAP$BMI + 
                        beta_age*dataPROPHYVAP$AGE, 
                      sd = 1)

#mean(cont_outcome)

dataPROPHYVAP$y_cont <- cont_outcome

.f <- formula(y_cont ~ GROUP_num * (AGE + SEX + ALCOOL + BMI + DIABETES))

### In practice use larger values of boot.number (e.g., 500).
### Here we set boot.number = 10 for speed in CRAN checks.
gc1 <- gc_continuous(formula=.f, model="all", data=dataPROPHYVAP, group="GROUP_num",
                    cv=10, boot.type="boot", boot.number=10, boot.tune=TRUE,
                    effect="ATE", progress=TRUE, seed=5192)

#Plot the calibration curve (by default)
plot(gc1, method="calibration", n.groups=5)
	 
#Plot the proportion comparison
plot(gc1, method="proportion")

}

\keyword{Plotting}
\keyword{G-computation}
\keyword{continuous}