\name{plot.gctimes}
\alias{plot.gctimes}
\title{

S3 Method for Plotting an 'gctimes' Object

}
\description{

Provides graphical summaries for an object returned by the function \code{gc_times}.
This S3 method offers two types of plots: a calibration plot and a survival curve plot.

}
\usage{

\method{plot}{gctimes}(x, method="calibration", n.groups=5,
                      pro.time=NULL, smooth=FALSE, ...)

}
\arguments{

\item{x}{An object returned by the function \code{gc_times}.}

\item{method}{A character string specifying the type of plot. Possible values are \code{calibration} (default) or \code{survival}.}

\item{n.groups}{An integer for the number of groups to divide the predicted probabilities into when \code{method="calibration"} is used. The default is 5.
Note: If \code{n.groups} is set too high, some groups may have too few events to reliably estimate survival. An error can also occur if the Q-model is singular.}

\item{pro.time}{The prognostic time at which the calibration plot of the survival probabilities is performed. If \code{NULL} (default), the median of the outcome time variable is used.}

\item{smooth}{A logical value by default equal to \code{FALSE}. If \code{boot.mi} is TRUE, enables plotting a single smooth calibration curve averaged over all \code{m} values instead of separate curves.}

\item{...}{Additional graphical parameters that can be passed to the underlying plot function.}

}
\details{

The \code{plot.gctimes} function is an S3 method for visualizing results from a \code{gctimes} object, which is generated by the \code{gc_times} function.

This plotting method supports two distinct types of visualizations:
\itemize{
\item \code{method="survival"}: This plot compares the overall Kaplan-Meier (KM) estimate of the survival curve from the original dataset against the mean of the survival predictions derived from the G-Computation model. The KM curve is typically displayed in blue, while the mean of the G-computation predictions is shown in red. This plot provides a visual assessment of how well the G-computation model's average survival predictions align with the observed survival experience in the raw data.

\item \code{method="calibration"}: This plot assesses the calibration of the Q-model's predictions used within G-Computation created on all of the raw data.
    \itemize{
    \item It takes the predicted survival probabilities at a specific \code{pro.time}. If \code{pro.time} is not specified, it defaults to the median of the outcome time variable.
    \item These predicted probabilities are then divided into \code{n.groups} (defaulting to 5) by their quantiles, creating different risk strata.
    \item For each group, the mean predicted survival probability is plotted on the X axe against the observed KM survival probability for that group at the specified \code{pro.time} on the Y axe.
    \item 95\% confidence intervals for these observed survival probabilities are also displayed as vertical segments.
    \item An identity line (y=x) is included for reference. Perfect calibration is indicated when observed points are along this line, suggesting that the predicted probabilities accurately reflect the observed survival rates across different risk groups.
    }
}


}
\value{No return value for this S3 method.}
\examples{
data(dataPROPHYVAP)

dataPROPHYVAP$DEATH_num <- ifelse(dataPROPHYVAP$DEATH == "Yes", 1, 0)
dataPROPHYVAP$GROUP_num <- ifelse(dataPROPHYVAP$GROUP == "Placebo", 0, 1)

formula <- formula(Surv(TIME_DEATH, DEATH_num) ~ GROUP_num * (AGE + SEX +
              BMI + DIABETES))

gc1 <- gc_times(formula=formula, model="lasso", data=dataPROPHYVAP,
              group="GROUP_num", param.tune=NULL, boot.type="bcv", cv=10,
              boot.number=10, effect="ATE", progress=TRUE , pro.time=10,
              boot.tune=TRUE)

plot(gc1)

#Plot the calibration curve
plot(gc1, method="calibration", n.groups=5)

#Plot the survival curves
plot(gc1, method="survival")

}
\keyword{Plotting}
\keyword{G-computation}
\keyword{times}
\keyword{Survival}